requirements.txt
arrow==1.3.0
asgiref==3.8.1
astroid==3.3.5
binaryornot==0.4.4
certifi==2024.12.14
chardet==5.2.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cookiecutter==2.6.0
dill==0.3.9
Django==5.1.3
django-browser-reload==1.17.0
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
iniconfig==2.0.0
isort==5.13.2
Jinja2==3.1.5
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mccabe==0.7.0
mdurl==0.1.2
packaging==24.2
pillow==11.0.0
platformdirs==4.3.6
pluggy==1.5.0
Pygments==2.18.0
PyJWT==2.10.1
pytest==8.3.3
pytest-django==4.9.0
python-dateutil==2.9.0.post0
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
six==1.17.0
sqlparse==0.5.2
text-unidecode==1.3
tomlkit==0.13.2
types-python-dateutil==2.9.0.20241206
tzdata==2024.2
Unidecode==1.3.8
urllib3==2.3.0
